multilang_TTS/requirements.txt
gtts
playsound
django
djangorestframework
django-cors-headers